agentic-sdr/requirements.txt
anthropic>=0.40.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-dotenv>=1.0.0
httpx>=0.27.0
rapidfuzz>=3.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
slack-sdk>=3.27.0
hubspot-api-client>=9.0.0
